% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{denom_lists}
\alias{denom_lists}
\title{Calculate denominator of Jaccard index for all combinations from two lists of character vectors}
\usage{
denom_lists(A, B)
}
\arguments{
\item{A}{list of character vectors}

\item{B}{list of character vectors}
}
\value{
a matrix of Jaccard indexes with rows corresponding to entries in A,
cols to entries in B
}
\description{
Calculate denominator of Jaccard index for all combinations from two lists of character vectors
}
