% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{denom_lists_wrapper}
\alias{denom_lists_wrapper}
\title{Calculate denominator of Jaccard index for two lists of character vectors}
\usage{
denom_lists_wrapper(A, B = NULL)
}
\arguments{
\item{A}{list of character vectors}

\item{B}{list of character vectors. If B is NULL, sets B=A.
(but denom_lists(A) is facter than denom_lists(A,A))}
}
\value{
a matrix of Jaccard indexes with rows corresponding to entries in A,
cols to entries in B
}
\description{
Calculate denominator of Jaccard index for two lists of character vectors
}
\author{
Chris Wallace
A=list(mammals=c("dog","cat"), others=c("fish","spider"))
B=list(domestic=c("cat","parrot","dog"), wild=c("lion","spider"))
denom_lists_wrapper(A,B)
}
