% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{denom_pair}
\alias{denom_pair}
\title{Calculate denominator of Jaccard index for two character vectors}
\usage{
denom_pair(x, y)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\value{
|union(x,y)|
}
\description{
Calculate denominator of Jaccard index for two character vectors
}
