% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{denom_symlist}
\alias{denom_symlist}
\title{Calculate denominator of Jaccard index between all combinations of a list of
character vectors with itself}
\usage{
denom_symlist(A)
}
\arguments{
\item{A}{list of character vectors}
}
\value{
a symmetric matrix of denominators of Jaccard indexes with rows and cols
corresponding to entries in A
}
\description{
Calculate denominator of Jaccard index between all combinations of a list of
character vectors with itself
}
