% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaccard_fast_matrix}
\alias{jaccard_fast_matrix}
\title{Get Jaccard simmilarity for matrix rows}
\usage{
jaccard_fast_matrix(mat)
}
\arguments{
\item{mat}{Binary m x n matrix to be evaluated}
}
\value{
Jaccard m x m simmilarity matrix
}
\description{
Get Jaccard simmilarity for matrix rows
}
