% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastJaccard-package.R, R/RcppExports.R
\name{jaccard_lists}
\alias{jaccard_lists}
\title{Calculate Jaccard index for two lists of character vectors}
\usage{
jaccard_lists(A, B)

jaccard_lists(A, B)
}
\arguments{
\item{A}{list of character vectors}

\item{B}{list of character vectors}
}
\value{
a matrix of Jaccard indexes with rows corresponding to entries in A,
cols to entries in B

a matrix of Jaccard indexes with rows corresponding to entries in A,
cols to entries in B
}
\description{
Calculate Jaccard index for two lists of character vectors

Calculate Jaccard index for two lists of character vectors
}
\author{
Chris Wallace
}
