% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastJaccard-package.R
\name{jaccard_lists_wrapper}
\alias{jaccard_lists_wrapper}
\title{Calculate Jaccard index for two lists of character vectors}
\usage{
jaccard_lists_wrapper(A, B = NULL)
}
\arguments{
\item{A}{list of character vectors}

\item{B}{list of character vectors. If B is NULL, sets B=A.
(but jaccard_lists(A) is facter than jaccard_lists(A,A))}
}
\value{
a matrix of Jaccard indexes with rows corresponding to entries in A,
cols to entries in B
}
\description{
Calculate Jaccard index for two lists of character vectors
}
\author{
Chris Wallace
}
