% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaccard_pair}
\alias{jaccard_pair}
\title{Calculate Jaccard index for two character vectors}
\usage{
jaccard_pair(x, y)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\value{
|intersect(x,y)| / |union(x,y)|
}
\description{
Calculate Jaccard index for two character vectors
}
