% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaccard_symlist}
\alias{jaccard_symlist}
\title{Calculate Jaccard indexes between all combinations of a list of character
vectors with itself}
\usage{
jaccard_symlist(A)
}
\arguments{
\item{A}{list of character vectors}
}
\value{
a symmetric matrix of Jaccard indexes with rows and cols
corresponding to entries in A
}
\description{
Calculate Jaccard indexes between all combinations of a list of character
vectors with itself
}
