% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overlap_lists}
\alias{overlap_lists}
\title{Calculate Jaccard index for two lists of character vectors}
\usage{
overlap_lists(A, B)
}
\arguments{
\item{A}{list of character vectors}

\item{B}{list of character vectors}
}
\value{
a matrix of count of overlaps with rows corresponding to entries in A,
cols to entries in B
}
\description{
Calculate Jaccard index for two lists of character vectors
}
