% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overlap_pair}
\alias{overlap_pair}
\title{Count overlap between two character vectors}
\usage{
overlap_pair(x, y)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\value{
|intersect(x,y)|
}
\description{
Count overlap between two character vectors
}
