% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overlap_symlist}
\alias{overlap_symlist}
\title{Count overlaps for a symmetric list of character vector}
\usage{
overlap_symlist(A)
}
\arguments{
\item{A}{list of character vectors}
}
\value{
a symmetric matrix of counts with rows and cols
corresponding to entries in A
}
\description{
Count overlaps for a symmetric list of character vector
}
