% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallelVectorSum}
\alias{parallelVectorSum}
\title{Parallel sum of a vector}
\usage{
parallelVectorSum(x)
}
\arguments{
\item{x}{Integer vector}
}
\value{
integer sum of vector
}
\description{
Parallel sum of a vector
}
