% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_jaccard_distance}
\alias{rcpp_parallel_jaccard_distance}
\title{Multiplies two doubles}
\usage{
rcpp_parallel_jaccard_distance(mat)
}
\arguments{
\item{mat}{Binary matrix to be evaluated}
}
\value{
Jaccard simmilarity matrix
}
\description{
Multiplies two doubles
}
