% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaccard_pair}
\alias{jaccard_pair}
\title{Calculate Jaccard index for two character vectors}
\usage{
jaccard_pair(x, y)
}
\arguments{
\item{x}{First vector, coerced to character if it is not}

\item{y}{Second vector, coerced to character if it is not}
}
\value{
Jaccard index |intersect(x,y)| / |union(x,y)|
}
\description{
Workhorse of the package, but will not be faster than a standard
jaccard for a single pair of vectors.
}
