% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{overlap_lists_wrapper}
\alias{overlap_lists_wrapper}
\title{Calculate overlap between two lists of character vectors}
\usage{
overlap_lists_wrapper(A, B = NULL)
}
\arguments{
\item{A}{list of character vectors}

\item{B}{list of character vectors. If B is NULL, sets B=A.
(but overlap_lists(A) is faster than overlap_lists(A,A))}
}
\value{
a matrix of overlap counts with rows corresponding to entries in A,
cols to entries in B
}
\description{
Calculate overlap between two lists of character vectors
}
\examples{
A=list(mammals=c("dog","cat"), others=c("fish","spider"))
B=list(domestic=c("cat","parrot","dog"), wild=c("lion","spider"))
overlap_lists_wrapper(A,B)
}
\author{
Chris Wallace
}
